adding calls to get which libraries does a user follow refactoring out some classes from library commander
