+++ kifi-backend/modules/shoebox/app/com/keepit/admin/WanderingAdminController.scala
@@ -14,7 +14,7 @@ import play.api.mvc.{ Result }
 import scala.concurrent.{ Future, Promise }
 import scala.util.{ Failure, Success }
 import scala.concurrent.duration._
-import com.keepit.commanders.LibraryInfo
+import com.keepit.model.LibraryInfo
 import com.keepit.social.BasicUser
 
 class WanderingAdminController @Inject() (
