+++ kifi-backend/modules/shoebox/app/com/keepit/common/curator/model/RecommendationInfo.scala
@@ -1,9 +1,8 @@
 package com.keepit.curator.model
 
-import com.keepit.commanders.FullLibraryInfo
 import com.keepit.common.crypto.PublicId
 import com.keepit.common.db.ExternalId
-import com.keepit.model.{ Library, NormalizedURI, URISummary }
+import com.keepit.model.{ FullLibraryInfo, Library, NormalizedURI, URISummary }
 import com.keepit.social.BasicUser
 import com.kifi.macros.json
 import org.joda.time.DateTime
