+++ kifi-backend/modules/shoebox/app/views/email/libraryInvitationText.scala.html
@@ -1,7 +1,7 @@
 @(inviteeId: Option[Id[User]],
 inviterId: Id[User],
 inviteMsg: Option[String],
-libraryInfo: com.keepit.commanders.LibraryInfo,
+libraryInfo: com.keepit.model.LibraryInfo,
 passPhrase: Option[String],
 authToken: String)
 @import com.keepit.common.mail.template.helpers._
